+++ coil-core/src/commonMain/kotlin/coil3/memory/RealMemoryCache.kt
@@ -35,7 +35,7 @@ internal class RealMemoryCache(
         return removedStrong || removedWeak
     }
 
-    override fun trimToSize(size: Long) {
+    override fun trimToSize(size: Int) {
         strongMemoryCache.trimToSize(size)
     }
