+++ coil-test/src/androidMain/kotlin/coil3/test/FakeImage.android.kt
@@ -16,7 +16,7 @@ import coil3.annotation.ExperimentalCoilApi
 actual class FakeImage actual constructor(
     actual override val width: Int,
     actual override val height: Int,
-    actual override val size: Long,
+    actual override val size: Int,
     actual override val shareable: Boolean,
     actual val color: Int,
 ) : Image {
