+++ coil-test/src/nonAndroidMain/kotlin/coil3/test/FakeImage.nonAndroid.kt
@@ -19,7 +19,7 @@ import org.jetbrains.skia.impl.use
 actual class FakeImage actual constructor(
     actual override val width: Int,
     actual override val height: Int,
-    actual override val size: Long,
+    actual override val size: Int,
     actual override val shareable: Boolean,
     actual val color: Int,
 ) : Image {
