+++ internal/test-utils/src/androidMain/kotlin/coil3/test/utils/FakeImage.android.kt
@@ -7,7 +7,7 @@ import coil3.Image
 actual class FakeImage actual constructor(
     actual override val width: Int,
     actual override val height: Int,
-    actual override val size: Long,
+    actual override val size: Int,
     actual override val shareable: Boolean,
 ) : Image {
     override fun asDrawable(resources: Resources): Drawable {
