+++ internal/test-utils/src/nonAndroidMain/kotlin/coil3/test/utils/FakeImage.nonAndroid.kt
@@ -6,7 +6,7 @@ import org.jetbrains.skia.Bitmap
 actual class FakeImage actual constructor(
     actual override val width: Int,
     actual override val height: Int,
-    actual override val size: Long,
+    actual override val size: Int,
     actual override val shareable: Boolean,
 ) : Image {
     override fun toBitmap(): Bitmap {
